% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.plot.R
\name{cdf.plot}
\alias{cdf.plot}
\title{Plot a discrete cumulative distribution function}
\usage{
cdf.plot(
  x,
  y,
  sort = FALSE,
  xlim,
  ylim = range(c(0, y)),
  xlab = "x",
  ylab = "F(x)",
  main = "",
  add = FALSE,
  verticals = TRUE,
  do.points = TRUE,
  pch = 20,
  col = 1,
  col.points = col,
  cex.points = 1,
  col.vert = col,
  lty = 1,
  lwd = 2,
  lty.vert = 3,
  lwd.vert = 1,
  ...
)
}
\arguments{
\item{x}{numeric vector giving the possible values of the discrete random variable.}

\item{y}{numeric vector giving the cumulative probabilities corresponding to \code{x}.}

\item{sort}{logical; if \code{FALSE} (default), the \code{x} values are assumed to be in
ascending order.}

\item{xlim, ylim}{the x and y limits of the plot.}

\item{xlab, ylab}{the axis titles.}

\item{main}{an overall title for the plot.}

\item{add}{logical; if \code{TRUE} add to an already existing plot, else generate a new plot.}

\item{verticals}{logical; if \code{TRUE}, draw vertical lines at steps.}

\item{do.points}{logical; if \code{TRUE} (default), draw points at the knot locations.}

\item{pch}{point character or symbol if \code{do.points}.}

\item{col}{default color of all points and lines.}

\item{col.points}{color of points if \code{do.points}.}

\item{cex.points}{symbol expansion factor if \code{do.points}.}

\item{col.vert}{color of vertical lines.}

\item{lty, lwd}{line type and thickness for horizontal lines.}

\item{lty.vert, lwd.vert}{line type and thickness for vertical lines.}

\item{...}{further graphical arguments (passed to \code{plot.default()} or to \code{segments()}).}
}
\value{
Returns a numeric vector with the random deviates and an attribute \code{ngen}
with the required number of uniform generations.
}
\description{
Plots a discrete/categorical cumulative distribution function.
}
\examples{
n <- 5
p <- 0.5
x <- 0:n
y <- pbinom(x, size = n, prob = p)
cdf.plot(x, y,
    main = paste0("B(", n,", ", p, ") cumulative distribution function"))
}
\seealso{
\code{\link[=plot.stepfun]{plot.stepfun()}}
}
