% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddexp.R
\name{ddexp}
\alias{ddexp}
\alias{pdexp}
\alias{qdexp}
\alias{rdexp}
\title{The double-exponential distribution}
\usage{
ddexp(x, rate = 1)

pdexp(q, rate = 1)

qdexp(p, rate = 1)

rdexp(n = 1000, rate = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{rate}{vector of rates.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}
}
\value{
\code{ddexp()} gives the density, \code{pdexp()} gives the distribution function,
\code{qdexp()} gives the quantile function, and \code{rdexp()} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the double-exponential distribution with rate \code{rate}.
}
\details{
If \code{rate} is not specified, it assumes the default value of 1.

The double-exponential distribution with rate \eqn{\lambda} has density
\deqn{f(x)  =\frac{\lambda}{2}e^{-\lambda | x |}}
for \eqn{x \in \mathbb{R}}.
The cummulative distribution is:
\deqn{F(x) = \int_{-\infty}^{x}f(t) dt
=\left\{
\begin{array}{ll}
\frac{1}{2}e^{\lambda x} & \text{if } x < 0\\
1-\frac{1}{2}e^{-\lambda x} & \text{if } x \geq 0
\end{array}
\ \right.}
The inverse cumulative distribution function is given by
\deqn{F^{-1}(p) = - \operatorname{sign}(p-0.5)\frac{\ln(1 - 2 | p - 0.5 | )}
{\lambda}.}
}
\examples{
set.seed(54321)
rate <- 2
rx <- rdexp(10^3, rate)
hist(rx, breaks = "FD", freq = FALSE)
lines(density(rx), lty = 2, col = "blue")
curve(ddexp(x, rate), lwd = 2, add = TRUE)
p <- c(0.005, 0.025, 0.05, 0.1)
abline(v = qdexp(c(p, 1-p), rate = rate), lty = 3, col = "red")
plot(ecdf(rx))
curve(pdexp(x, rate), lwd = 2, add = TRUE)
}
\seealso{
\code{\link[=dexp]{dexp()}}
}
