% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{freq.test}
\alias{freq.test}
\title{Frequency test}
\usage{
freq.test(u, nclass = floor(length(u)/10))
}
\arguments{
\item{u}{numeric vector containing the observed values
(or the pseudorandom numbers).}

\item{nclass}{number of bins (to partition the unit interval).}
}
\value{
A list with class \code{"htest"} returned by \code{\link[=chisq.test]{chisq.test()}}.
}
\description{
Performs the chi-squared goodness-of-fit test for the particular case
of an uniform distribution in the unit interval
(by grouping data into \emph{bins}, \code{x = floor(nclass*u) + 1},
and calling \code{\link{chisq.test}}\code{(x)}).
Used for testing random number generators.
}
\examples{
set.rng(321, "lcg", a = 5, c = 1, m = 512)    # set.seed(321)
nsim <- 500
res <- freq.test(rng(nsim))
res # Pearson's Chi-squared test
# Pearson's Chi-squared test table
sapply(res[c("observed", "expected", "residuals", "stdres")], as.vector)
# Plot observed and expected counts.
plot(res$observed)
abline(h = res$expected[1], lty = 2)
}
\seealso{
\code{\link{chisq.cont.test}} \code{\link{chisq.test}}
}
