% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simres_package.R
\docType{data}
\name{lifetimes}
\alias{lifetimes}
\title{Microorganism lifetimes}
\format{
A vector with 15 observations of the time to death event.
}
\source{
Unknown.
}
\description{
Lifespan of a sample of a particular population of microorganisms.
}
\examples{
hist(lifetimes)
}
\keyword{datasets}
