% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{rcauchy.rou}
\alias{rcauchy.rou}
\title{Ratio-of-uniforms method for Cauchy distribution generation}
\usage{
rcauchy.rou(n)
}
\arguments{
\item{n}{number of observations.}
}
\value{
Returns a numeric vector with the random deviates and an attribute \code{ngen}
with the required number of uniform generations.
}
\description{
Uses the ratio-of-uniforms method for simulating a standard Cauchy distribution.
}
\examples{
set.seed(1)
nsim <- 10^4
rx <- rcauchy.rou(nsim)
hist(rx, breaks = "FD", freq = FALSE, main = "", xlim = c(-6, 6))
curve(dcauchy, add = TRUE)
ngen <- attr(rx, "ngen")
{cat("Number of generations = ", ngen)
  cat("\nAveraged number of generations = ", ngen/nsim)
  cat("\nRejection rate = ", 1-nsim/ngen,"\n")}
}
\seealso{
\code{\link[=rcauchy]{rcauchy()}}
}
