% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{rephtest}
\alias{rephtest}
\alias{summary.rhtest}
\alias{plot.rhtest}
\title{Test repetition}
\usage{
rephtest(n = 30, nsim = 1000, test, rand.gen = runif, ...)

\method{summary}{rhtest}(object, alpha = c(0.01, 0.05, 0.1, 0.25, 0.5), ...)

\method{plot}{rhtest}(x, y = 1:3, ask = length(y) > 1 && dev.interactive(), ...)
}
\arguments{
\item{n}{sample size.}

\item{nsim}{number of simulations.}

\item{test}{function (or function name) which performs an one-sample test.}

\item{rand.gen}{optional: function to generate the samples.}

\item{...}{arguments to be passed to other functions (for instance to \code{test()})
or methods.}

\item{object}{an object for which a summary is desired.}

\item{alpha}{numeric vector of probabilities (significance levels).}

\item{x}{an object with class \code{"rhtest"}.}

\item{y}{if a subset of the plots is required,
specify a subset of the numbers \code{1:3}.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot,
see \code{\link[graphics:par]{graphics::par}}\code{(ask=.)}.}
}
\value{
\code{rephtest()} returns a list with class \code{"rhtest"} containing the following components:
\item{statistics}{the values of the test statistic.}
\item{p.values}{the p-values for the test.}
with attributes:
\item{method}{a character string indicating the type of test performed.}
\item{names.stat}{a character string indicating the distribution
of the test statistic.}
\item{parameter}{the parameters of the distribution of the test statistic.}
}
\description{
Applies a hypothesis test to simulated samples.
}
\examples{
set.rng(543210, "lcg", a = 2^16 + 3, c = 0, m = 2^31)  # set.seed(543210)
res <- rephtest(n = 30, test = chisq.cont.test, rand.gen = rng,
                 distribution = "unif", output = FALSE, nestpar = 0)
str(res)
summary(res)
plot(res)
}
\seealso{
\code{\link{chisq.cont.test}} \code{\link{freq.test}}
}
