% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmf.R
\name{rpmf}
\alias{rpmf}
\title{Inverse transform method for discrete distribution generation}
\usage{
rpmf(x, prob = 1/length(x), n = 1000, as.factor = FALSE)
}
\arguments{
\item{x}{numeric vector giving the possible values of the discrete random variable.}

\item{prob}{numeric vector giving the probabilities corresponding to \code{x}.}

\item{n}{number of observations to generate.}

\item{as.factor}{logical; if \code{TRUE}, the returned vector is encoded as a factor
with levels \code{x}.}
}
\value{
Returns a numeric vector, or a factor if \code{as.factor = TRUE}, with the random
deviates and an attribute \code{ncomp} with the required number of comparisons in
the sequential search.
}
\description{
Uses the inverse transform method with sequential search for simulating a
discrete/categorical probability distribution (that takes on only a finite
number of values) from its probability mass function.
}
\examples{
set.seed(1)
# Simulation of a binomial distribution
n <- 10
p <- 0.5
nsim <- 10^5
x <- 0:n
pmf <- dbinom(x, n, p)
rx <- rpmf(x, pmf, nsim)
# Relative frequency plot
plot(table(rx)/nsim, ylab = "Relative frequency", xlab = "Value")
abline(v = mean(rx))
# Theoretical values
points(x, pmf, pch = 4, col = "blue")
abline(v = p*n, lty = 2, col = "blue")
# Number of comparisons
ncomp <- attr(rx, "ncomp")
ncomp/nsim # mean number
sum((1:length(x))*pmf) # theoretical expected number
}
\seealso{
\code{\link[=sample]{sample()}}, \code{\link[=rpmf.table]{rpmf.table()}}, \code{\link[=rpmf.alias]{rpmf.alias()}}
}
