% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpmf.R
\name{rpmf.table}
\alias{rpmf.table}
\title{Guide table method for discrete distribution generation}
\usage{
rpmf.table(x, prob = 1/length(x), m, n = 1000, as.factor = FALSE)
}
\arguments{
\item{x}{numeric vector giving the possible values of the discrete random variable.}

\item{prob}{numeric vector giving the probabilities corresponding to \code{x}.}

\item{m}{size of the table with starting points for the search.}

\item{n}{number of observations to generate.}

\item{as.factor}{logical; if \code{TRUE}, the returned vector is encoded as a factor
with levels \code{x}.}
}
\value{
Returns a numeric vector, or a factor if \code{as.factor = TRUE}, with the random
deviates and an attribute \code{ncomp} with the required number of comparisons in
the sequential search.
}
\description{
Uses the guide table aided inversion method (also know as indexed search method)
for simulating a discrete/categorical probability distribution (that takes on
only a finite number of values) from its probability mass function.
}
\examples{
set.seed(1)
# Simulation of a binomial distribution
n <- 10
p <- 0.5
nsim <- 10^5
x <- 0:n
pmf <- dbinom(x, n, p)
rx <- rpmf.table(x, pmf, n-1, nsim)
# Relative frequency plot
plot(table(rx)/nsim, ylab = "Relative frequency", xlab = "Value")
abline(v = mean(rx))
# Theoretical values
points(x, pmf, pch = 4, col = "blue")
abline(v = p*n, lty = 2, col = "blue")
# Number of comparisons
ncomp <- attr(rx, "ncomp")
ncomp/nsim # mean number
sum((1:length(x))*pmf) # theoretical expected number with sequential search
}
\seealso{
\code{\link[=sample]{sample()}}, \code{\link[=rpmf]{rpmf()}}, \code{\link[=rpmf.alias]{rpmf.alias()}}
}
