% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.time.R, R/simres-internals.R, R/test.R
\name{.cpu.time.ini}
\alias{.cpu.time.ini}
\alias{simres-internals}
\alias{eval.curve}
\alias{print.summaryrhtest}
\title{simres internal and secondary functions}
\usage{
.cpu.time.ini()

eval.curve(expr, x, enclos = parent.frame())

\method{print}{summaryrhtest}(x, ...)
}
\arguments{
\item{expr}{name of the function (or call or expression written as a function)
of \code{x} which will evaluate to an object of the same length as \code{x}.
See for instance \code{\link{curve}}.}

\item{enclos}{an \emph{enclosure} where looks for objects not found in \code{envir}.}

\item{envir}{the environment in which \code{expr} is to be evaluated.}
}
\description{
Listed below are supporting functions for the major methods in simres.

\code{eval.curve} evaluates the curve (or expression) \code{expr} in the environment
specified by \code{enclos}.
}
\keyword{internal}
